requirements.txt
datasets==4.8.4
fastapi==0.135.3
medmnist==3.0.2
numpy==2.4.4
Pillow==12.2.0
scikit_learn==1.8.0
scipy==1.17.1
torch==2.11.0
transformers==5.4.0
uvicorn==0.42.0
